create olScript utility module
