+++ lib/mapp.mjs
@@ -34,21 +34,12 @@ import plugins from './plugins/_plugins.mjs'
 hooks.parse();
 
 const _ol = {
-  current: '10.2.1',
-  load: async () => await new Promise(resolve => {
+  current: '10.2.1'
+}
 
-    const script = document.createElement('script')
+async function olLoad() {
 
-    script.type = 'application/javascript'
 
-    script.src = 'https://cdn.jsdelivr.net/npm/ol@v10.2.1/dist/ol.js'
-
-    script.onload = resolve
-
-    document.head.append(script)
-
-    console.warn('Openlayers library loaded from script tag.')
-  })
 }
 
 if (window.ol === undefined) {
