+++ lib/utils/_utils.mjs
@@ -34,6 +34,10 @@ import getCurrentPosition from './getCurrentPosition.mjs'
 
 import merge from './merge.mjs'
 
+import mobile from './mobile.mjs';
+
+import olScript from './olScript.mjs'
+
 import paramString from './paramString.mjs'
 
 import { polygonIntersectFeatures } from './polygonIntersectFeatures.mjs'
@@ -60,8 +64,6 @@ import { formatNumericValue, unformatStringValue } from './numericFormatter.mjs'
 
 import { versionCheck } from './versionCheck.mjs';
 
-import mobile from './mobile.mjs';
-
 export default {
   stats,
   render,
@@ -81,7 +83,7 @@ export default {
   loadPlugins,
   merge,
   mobile,
-  olScript: () => mapp.ol.load(),
+  olScript,
   paramString,
   polygonIntersectFeatures,
   promiseAll,
